Shellcode Assignment
We did a shellcode assignment in Assembly. There was a function named "bof" on our class's putty server in which we had to get the keyword that was in a folder in the directory that the program was in. We were given shellcode and we had to edit the code to remove NULL bytes and reduce the size.